feat(camera) :sparkles: added camera values for device

+++ app/src/main/java/com/iiap/deeplarva/domain/response/CameraConfigurationResponse.kt
@@ -3,5 +3,5 @@ package com.iiap.deeplarva.domain.response
 data class CameraConfigurationResponse (
     val iso: Int,
     val exposure: Float,
-    val shutterSpeed: Long,
+    val shutterSpeed: Int,
 )
